Split certificate management into Authify.Certificates context
Extract certificate CRUD, generation, and active-certificate lookup
functions from the Accounts context into a dedicated Authify.Certificates
context module. Update all callers and tests to use the new context.
